requirements.txt
pyOpenSSL
flask
pytest
flask-login
duo-web
requests
werkzeug
Flask-Mail
oauthlib
psycopg2
gunicorn
python-dotenv
virtualenv
Flask-SQLAlchemy
Flask-Migrate
Flask-Script